Fix: Restore missing Change-Id commit (#86)
* refactor: User Entity id 수정

* refactor: User Id 변경에 따른 CustomOAuth2UserService 리팩토링

* refactor: User 엔티티 name 필드 삭제

* refactor: User ID 변경에 따른 OAuth2User 구현체 리팩토링

* refactor: OAuth2UserService 리팩토링

* refactor: CustomOAuth2UserService 트랜잭션 범위 변경

* refactor: GithubUserProfileDto 이름 변경

* refactor: UserId 변경에 따른 JWT 관련 로직 수정

* refactor: batch 패키지 경로 변경

* refactor: UserId 변경 Ranking Batch 작업에 반영

* refactor: UserId 변경 Ranking, Follow 도메인에 반영

+++ src/main/java/com/moyo/backend/domain/github_ranking/data_access/RankingRepositoryImpl.java
@@ -46,7 +46,7 @@ public class RankingRepositoryImpl implements RankingRepository {
 	}
 
 	@Override
-	public Slice<Ranking> findFollowingUserRankings(List<Long> followingUserIds, RankingPeriod rankingPeriod, Pageable pageable) {
+	public Slice<Ranking> findFollowingUserRankings(List<Integer> followingUserIds, RankingPeriod rankingPeriod, Pageable pageable) {
 		return rankingQueryDslRepository.findByUserIds(followingUserIds, rankingPeriod, pageable);
 	}
 }
